requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0.1
django-debug-toolbar==4.3.0
django-filter==23.5
djangorestframework==3.14.0
filelock==3.13.1
idna==3.6
Markdown==3.5.2
mysqlclient==2.2.1
platformdirs==4.2.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
virtualenv==20.25.0
